Use KeyMap instead of HashMap when dealing with Objects
In `aeson-2.0.*`, `Object`'s field is of type `KeyMap Value` rather than
`HashMap Text Value`. Now that Stackage Nightly includes `aeson-2.0.*`, it's
high time that we updated the `lens-aeson` API to mirror this change. This most
notably affects the type of `_Object`, which also requires updating the types
of `key` and `members` as a consequence.

In order to make the doctests continue to work, I found myself needing to add
`Wrapped` and `Rewrapped` instances for `KeyMap`. In these instances, I treat
`KeyMap v` as a wrapper around `[(Key, v)]`, where the order of the key-value
pairs in the list is not stable.

Fixes #38.